requirements.txt
requests
beautifulsoup4
feedparser
python-dotenv
openai
tqdm
# optional
playwright
